requirements.txt
absl-py==2.0.0
accelerate==0.20.3
annotated-types==0.5.0
blessed==1.20.0
brotlipy==0.7.0
cachetools==5.3.2
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.6
coloredlogs==15.0.1
cryptography==39.0.1
cycler==0.11.0
deepspeed==0.12.6
filelock==3.12.2
flit_core==3.6.0
fonttools==4.38.0
fsspec==2023.1.0
google-auth==2.23.4
google-auth-oauthlib==0.4.6
gpustat==1.1
grpcio==1.59.2
h5py==3.8.0
hjson==3.1.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.7.0
joblib==1.3.1
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
munkres==1.1.4
ninja==1.11.1.1
nltk==3.8.1
numpy==1.19.2
nvidia-ml-py==12.535.77
oauthlib==3.2.2
packaging==23.1
Pillow==9.0.1
pip==22.3.1
protobuf==3.20.3
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pynvml==11.5.0
pyOpenSSL==23.0.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.6.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
setuptools==65.6.3
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2
thop==0.1.1.post2209072238
tokenizers==0.13.3
torch==1.13.0+cu117
torchaudio==0.13.1
torchprofile==0.0.4
torchvision==0.14.1
tornado==6.1
tqdm==4.65.0
transformers==4.30.2
typing_extensions==4.7.1
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.3
wheel==0.38.4
zipp==3.15.0
